# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ForbiddenError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UUID": {
				"type": "string",
				"example": "52907745-7672-470e-a803-a2f8feb52944",
				"format": "uuid",
				"description": "Stringified UUIDv4.\r\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
...
			"DefaultString": {
				"type": "string",
				"example": "string",
				"minLength": 2,
				"maxLength": 50
			},
...
			"BacklogCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"branchId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UUID"
							}
						],
						"nullable": true
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/CategoryCompleteOut"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"branchId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BranchNotFound": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryNameCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"name": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DefaultString"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"SessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session-id"
			},
			"JWTAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "access-token"
			}
		}
	},
	"info": {
		"title": "Menuchi-API",
		"version": "1.0.0",
		"description": "Menuchi API Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Mohammad Hosseini"
		}
	},
	"paths": {
...
		"/menus/backlog/{backlogId}": {
			"get": {
				"operationId": "GetBacklog",
				"responses": {
					"200": {
						"description": "Backlog is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4049 BranchNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BranchNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a backlog by ID.",
				"tags": [
					"Menu"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/DefaultString"
						}
					}
				]
			}
		},
...
		"/category-names": {
			"post": {
				"operationId": "CreateCategoryName",
				"responses": {
					"201": {
						"description": "Category name created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryNameCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError -> A category name with the provided name already exists.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					},
					"422": {
						"description": "4222 CategoryNameValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryNameValidationError"
								}
							}
						}
					}
				},
				"description": "Creates a new category name.",
				"tags": [
					"Category Name"
				],
				"security": [
					{
						"": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryNameCompactIn"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllCategoryNames",
				"responses": {
					"200": {
						"description": "All category names retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryNameCompleteOut"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Retrieves all category names.",
				"tags": [
					"Category Name"
				],
				"security": [
					{
						"": [
							"ADMIN",
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": []
			}
		},
...
		"/backlog/{backlogId}/items": {
			"post": {
				"operationId": "CreateItem",
				"responses": {
					"201": {
						"description": "Item created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateItemCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4044 BacklogNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogNotFound"
								}
							}
						}
					},
					"422": {
						"description": "4223 ItemValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemValidationError"
								}
							}
						}
					}
				},
				"description": "Creates a new backlog item and a category (if needed), using an existing category name.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ItemCompactIn"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetItems",
				"responses": {
					"200": {
						"description": "All backlog items retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ItemCompleteOut"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Retrieves all items in a backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteItems",
				"responses": {
					"204": {
						"description": "Items deleted successfully. It doesn't retrieve anything.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes multiple items from a backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/backlog/{backlogId}": {
			"get": {
				"operationId": "GetBacklog",
				"responses": {
					"200": {
						"description": "Backlog is retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4044 BacklogNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BacklogNotFound"
								}
							}
						}
					}
				},
				"description": "Retrieves a specific backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/backlog/{backlogId}/items/{itemId}": {
			"patch": {
				"operationId": "UpdateItem",
				"responses": {
					"204": {
						"description": "Item updated successfully. It doesn't retrieve anything.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "4223 ItemValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemValidationError"
								}
							}
						}
					}
				},
				"description": "Updates an item in a backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateItemIn"
							}
						}
					}
				}
			}
		},
		"/backlog/{backlogId}/reorder-items/in-category": {
			"patch": {
				"operationId": "ReorderItemsInCategory",
				"responses": {
					"204": {
						"description": "Item orders in the category updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "Some item IDs are invalid or do not belong to the backlog.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders items within a category.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/backlog/{backlogId}/reorder-items/in-list": {
			"patch": {
				"operationId": "ReorderItemsInList",
				"responses": {
					"204": {
						"description": "Item orders in the list updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "Some item IDs are invalid or do not belong to the backlog.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders items within a list.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/backlog/{backlogId}/categories": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Category created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryCompactOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError -> This category name is already used in this backlog.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					}
				},
				"description": "Creates a new category. Alternatively, creating an item will automatically create its category (if needed).",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryCompactIn"
							}
						}
					}
				}
			}
		},
		"/backlog/{backlogId}/reorder-categories": {
			"patch": {
				"operationId": "ReorderCategoriesInBacklog",
				"responses": {
					"204": {
						"description": "Category orders in the backlog updated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "All category IDs must be in the request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuchiError"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Reorders categories within a backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/backlog/{backlogId}/categories/{categoryId}": {
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"204": {
						"description": "Category and its items deleted successfully.  It doesn't retrieve anything.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes a category and all its items.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "path",
						"name": "categoryId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
		"/backlog/{backlogId}/category-names": {
			"get": {
				"operationId": "GetAllCategoryNames",
				"responses": {
					"200": {
						"description": "All category names retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryNameCompleteOut"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Retrieves all category names used in the specific backlog.",
				"tags": [
					"Backlog"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "backlogId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}